% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafusion-sql.R
\name{simulate_datafusion}
\alias{simulate_datafusion}
\alias{sql_translation.DatafusionConnection}
\title{Datafusion SQL Translation}
\usage{
simulate_datafusion()

sql_translation.DatafusionConnection(con)
}
\value{
\itemize{
\item \code{simulate_datafusion()} returns an object of class "DatafusionConnection"
that can be used to test SQL generation
\item \code{sql_translation.DatafusionConnection()} returns a \code{\link[dbplyr:sql_variant]{dbplyr::sql_variant()}}
that can be used to translate R expressions into SQL.
}
}
\description{
Datafusion SQL Translation
}
\examples{
simulate_datafusion()


}
