% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafusion-package.R
\docType{package}
\name{datafusion-package}
\alias{datafusion}
\alias{datafusion-package}
\title{datafusion: Experimental R Bindings to Datafusion}
\description{
Provides and experimental R interface to the 'DataFusion' query engine.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
