% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafusion-sql.R
\name{simulate_datafusion}
\alias{simulate_datafusion}
\title{Datafusion SQL Translation}
\usage{
simulate_datafusion()
}
\value{
\itemize{
\item \code{simulate_datafusion()} returns an object of class "DatafusionConnection"
that can be used to test SQL generation
\item \code{sql_translation.DatafusionConnection()} returns a \code{\link[dbplyr:sql_variant]{dbplyr::sql_variant()}}
that can be used to translate R expressions into SQL.
}
}
\description{
Datafusion SQL Translation
}
\examples{
library(dplyr, warn.conflicts = FALSE)

lf <- dbplyr::lazy_frame(a = TRUE, b = 1, c = 2, d = "z", con = simulate_datafusion())
lf \%>\% summarise(x = sd(b, na.rm = TRUE))
lf \%>\% summarise(y = cor(b, c), z = cov(b, c))

}
